last methods from user service refactored
