+++ components/app/myrw/areas/pages/alerts.js
@@ -12,8 +12,7 @@ import { getLabel } from 'utils/datasets/dataset-helpers';
 import AlertWidget from 'components/areas/AlertWidget';
 
 // Services
-import UserService from 'services/user';
-import { getSubscriptions } from 'services/subscriptions';
+import { getSubscription } from 'services/subscriptions';
 
 class AreasAlerts extends React.Component {
   constructor(props) {
@@ -23,9 +22,8 @@ class AreasAlerts extends React.Component {
     const { subscription } = user.areas.items.find(alert => alert.id === id);
 
     this.state = { subscriptionData: null };
-    this.userService = new UserService({ apiURL: process.env.CONTROL_TOWER_URL });
 
-    this.userService.getSubscription(subscription.id, user.token).then(((data) => {
+    getSubscription(subscription.id, user.token).then(((data) => {
       this.setState({ subscriptionData: data });
     }));
   }
