+++ components/subscriptions/SubscriptionCard.js
@@ -13,7 +13,7 @@ import Map from 'components/widgets/map/Map';
 // Services
 import DatasetService from 'services/DatasetService';
 import { getCountry, getGeostore } from 'services/areas';
-import UserService from 'services/user';
+import { deleteSubscription } from 'services/subscriptions';
 
 // Utils
 import LayerManager from 'utils/layers/LayerManager';
@@ -43,12 +43,6 @@ class SubscriptionCard extends React.Component {
       apiURL: process.env.WRI_API_URL,
       language: props.locale
     });
-    this.userService = new UserService({ apiURL: process.env.WRI_API_URL });
-
-    // ------------------- Bindings ---------------------------
-    this.handleDeleteSubscription = this.handleDeleteSubscription.bind(this);
-    this.handleGoToDataset = this.handleGoToDataset.bind(this);
-    // --------------------------------------------------------
   }
 
   componentDidMount() {
@@ -125,12 +119,12 @@ class SubscriptionCard extends React.Component {
       .catch(err => toastr.error('Error', err));
   }
 
-  handleDeleteSubscription() {
+  handleDeleteSubscription = () => {
     const { subscription, token } = this.props;
     const toastrConfirmOptions = {
       onOk: () => {
         this.setState({ loading: true });
-        this.userService.deleteSubscription(subscription.id, token)
+        deleteSubscription(subscription.id, token)
           .then(() => {
             this.props.onSubscriptionRemoved();
           })
@@ -140,7 +134,7 @@ class SubscriptionCard extends React.Component {
     toastr.confirm('Are you sure you want to delete the subscription?', toastrConfirmOptions);
   }
 
-  handleGoToDataset() {
+  handleGoToDataset = () => {
     Router.pushRoute('explore_detail', { id: this.props.subscription.attributes.datasets[0] });
   }
